% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.r
\name{animPlot}
\alias{animPlot}
\title{Animation plot}
\usage{
animPlot(data, ccm, delay = 0.1, darken = 1, lwd = 1, c.cex = 1,
  main = "", ...)
}
\arguments{
\item{data}{Dimensionality reduction plot}

\item{ccm}{Dataframe of indices and momentums}

\item{delay}{Delay between frames in seconds}

\item{darken}{Passed to colGrad() function}

\item{lwd}{Line width}

\item{c.cex}{Size of poiints.}

\item{main}{Plot title}

\item{...}{Passed to plot() function}
}
\value{
Generates plot
}
\description{
Animation plot
}
\details{
Animation plot
Generates plot in base R that gradually updates giving the impression of an animation
}
\examples{
\dontrun{
xx <- animPlot(x, ccm)
}
}
\author{
Wajid Jawaid
}
