% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.r
\name{animPlotGif}
\alias{animPlotGif}
\title{Generates a GIF animating}
\usage{
animPlotGif(data, ccm, delay = 0.1, darken = 1, lwd = 1, c.cex = 1,
  main = "", gif = "animation", img.name = "tempPlot", plot.par = NULL,
  point.col = "#333333", arrowLength = 0.1, ...)
}
\arguments{
\item{data}{Reduced dimensionality map to be used for visualisation}

\item{ccm}{Dataframe of indices and momentums}

\item{delay}{Delay between frames in seconds}

\item{darken}{Passed to colGrad() function}

\item{lwd}{Line width}

\item{c.cex}{Size of poiints.}

\item{main}{Title}

\item{gif}{Name of movie}

\item{img.name}{Name of temporary image files generated}

\item{plot.par}{Passed to R base par() function}

\item{point.col}{Colour of background points}

\item{arrowLength}{Modify length of arrow}

\item{...}{Passed to plot() function}
}
\value{
Produces an animated GIF with given file name
}
\description{
Generates a GIF animation
}
\details{
Generates a GIF animation
}
\examples{
\dontrun{
xx <- animPlotGif(x, ccm, gif = "animation")
}
}
\author{
Wajid Jawaid
}
