% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.r
\name{calculateVariableSigmas}
\alias{calculateVariableSigmas}
\title{Calculates sigmas for a distance matrix}
\usage{
calculateVariableSigmas(d, knn)
}
\arguments{
\item{d}{Square distance matrix with 0 diagonal}

\item{knn}{Number of nearest neighbours to use for calculation}
}
\value{
Returns a vector of sigmas
}
\description{
Calculates sigmas for a distance matrix
}
\details{
Calculates sigmas for a distance matrix
Using Laleh Hagherverdi's method
}
\examples{
\dontrun{
sigmas <- calculateVariableSigmas(dist, 5)
}
}
\author{
wj241
}
