% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{colGrad}
\alias{colGrad}
\title{Generates a smooth colour gradient}
\usage{
colGrad(x, darken = 1)
}
\arguments{
\item{x}{Number of colours required}

\item{darken}{Multiplication factor. Must be less than 1. Smaller the darker.}
}
\value{
Returns vector of RGB colours
}
\description{
Generates a smooth colour gradient
}
\details{
Generates a smooth colour gradient
Goes from red to red/green to green to green/blue to blue to blu/red
}
\examples{
gradientColors <- colGrad(10)
}
\author{
Wajid Jawaid
}
