% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.r
\name{diffuseProj}
\alias{diffuseProj}
\title{Predicts diffusion map projection from new data points}
\usage{
diffuseProj(dm, x, data, distfun)
}
\arguments{
\item{dm}{Output from diffuseMat2 function}

\item{x}{Matrix of new data points. Features in rows and cells in
columns.}

\item{data}{Original data used to generate diffusion map}

\item{distfun}{A distance function that takes new data as first
paramter and previous data as second variable returning a squared
distance measure, with each sample in the rows and distance to
previous data points in columns, e.g. function(x, y) (1 - cor(x, y))^2.}
}
\value{
Returns a matrix with projected diffusion components.
}
\description{
Predicts diffusion map projection from new data points
}
\details{
Predicts diffusion map projection from new data points
}
\examples{
\dontrun{
y <- diffuseProj(xx, newData, oldData, function(z) (1-cor(z))^2)
}
}
\author{
Wajid Jawaid
}
