% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt.r
\name{dpt}
\alias{dpt}
\title{Calculate pseudotime}
\usage{
dpt(dmap, startIndex)
}
\arguments{
\item{dmap}{Output from diffuseMat function}

\item{startIndex}{Numerical index of starting cell}
}
\value{
Vector of pseudotimes
}
\description{
Calculate pseudotime
}
\details{
Calculate diffusion pseudotime
}
\author{
Wajid Jawaid
}
