% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.r
\name{fastDist}
\alias{fastDist}
\title{Fast vectorised Euclidean distance calculator}
\usage{
fastDist(x, squared = FALSE)
}
\arguments{
\item{x}{Matrix with vectors in columns.}

\item{squared}{Will not perform the square root, i.e. will return the squared `L2-norm'.}
}
\value{
Returns a matrix of pairwise distances
}
\description{
Fast vectorised Euclidean distance calculator
}
\details{
Calculates Euclidean distances between vectors arranged as columns in a matrix.
}
\examples{
\dontrun{
dist <- fastDist(x)
}
}
\author{
Wajid Jawaid
}
