% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes}
\usage{
filterGenes(x, mu = 0.01, cv = 2, fano = FALSE)
}
\arguments{
\item{x}{Matrix to be normalised with cells in rows and
genes in columns}

\item{mu}{Meam threshold}

\item{cv}{Coefficient of variation or Fano factor threshold.}

\item{fano}{Default TRUE. Predicate treat CV as Fano factor or CV}
}
\value{
Returns a filtered matrix with same number of cells but fewer
genes than 'x'
}
\description{
Filter genes
}
\details{
Filter genes
Filter genes by mean and either coefficient of variation, cv or
Fano factor.
}
\examples{
\dontrun{
expressionGenesFiltered <- filterGenes(x)
}
}
\author{
Wajid Jawaid
}
