% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust.r
\name{findLouvain}
\alias{findLouvain}
\title{Louvain clustering on transition matrix}
\usage{
findLouvain(mkv)
}
\arguments{
\item{mkv}{Transition matrix}
}
\value{
Returns a list with graph, dataframe and community object
}
\description{
Louvain clustering on transition matrix
}
\details{
Louvain clustering on transition matrix
}
\examples{
\dontrun{
xx <- findLouvain(mkv)
xx$cll
}
}
\author{
Wajid Jawaid
}
