% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.r
\name{sparseMarkov}
\alias{sparseMarkov}
\title{Make markov matrix sparse}
\usage{
sparseMarkov(mkv, knn)
}
\arguments{
\item{mkv}{Markov matric}

\item{knn}{Number of nearest neighbours. See above.}
}
\value{
Markovian sparse matrix.
}
\description{
Make markov matrix sparse
}
\details{
Make markov matrix sparse
Choose knn as the maximum number of similar cells are likely to exist in your dataset.
}
\author{
Wajid Jawaid
}
